Added a mountain track
added a new track which is currently my favourite. I think I will end up using the heightmap + procedural geometry track as a baseline for the rest of the tracks I make for this game. Adding some props and affectors to it will make it a more complete map. I just need to reduce the ability to use "skips"